+++ src/solidlsp/language_servers/elm_language_server.py
@@ -151,7 +151,7 @@ class ElmLanguageServer(SolidLanguageServer):
 
         self.server.on_notification("window/logMessage", window_log_message)
         self.server.on_notification("$/progress", do_nothing)
-        self.server.on_notification("textDocument/publishDiagnostics", do_nothing)
+        self.server.on_notification("textDocument/publishDiagnostics", self._handle_publish_diagnostics)
 
         log.info("Starting Elm server process")
         self.server.start()