+++ src/solidlsp/language_servers/haskell_language_server.py
@@ -359,7 +359,7 @@ class HaskellLanguageServer(SolidLanguageServer):
 
         self.server.on_notification("window/logMessage", window_log_message)
         self.server.on_notification("$/progress", do_nothing)
-        self.server.on_notification("textDocument/publishDiagnostics", do_nothing)
+        self.server.on_notification("textDocument/publishDiagnostics", self._handle_publish_diagnostics)
         self.server.on_request("client/registerCapability", register_capability_handler)
         self.server.on_request("workspace/configuration", workspace_configuration_handler)