+++ src/solidlsp/language_servers/powershell_language_server.py
@@ -296,7 +296,7 @@ class PowerShellLanguageServer(SolidLanguageServer):
         self.server.on_request("client/registerCapability", register_capability_handler)
         self.server.on_notification("window/logMessage", window_log_message)
         self.server.on_notification("$/progress", do_nothing)
-        self.server.on_notification("textDocument/publishDiagnostics", do_nothing)
+        self.server.on_notification("textDocument/publishDiagnostics", self._handle_publish_diagnostics)
         self.server.on_notification("powerShell/executionStatusChanged", do_nothing)
 
         log.info("Starting PowerShell Editor Services process")