+++ src/solidlsp/language_servers/vue_language_server.py
@@ -669,7 +669,7 @@ class VueLanguageServer(SolidLanguageServer):
         self.server.on_notification("tsserver/request", tsserver_request_notification_handler)
         self.server.on_notification("window/logMessage", window_log_message)
         self.server.on_notification("$/progress", do_nothing)
-        self.server.on_notification("textDocument/publishDiagnostics", do_nothing)
+        self.server.on_notification("textDocument/publishDiagnostics", self._handle_publish_diagnostics)
 
         log.info("Starting Vue server process")
         self.server.start()